+++ redis/lib/testcontainers/redis.rb
@@ -13,18 +13,14 @@ module Testcontainers
     # Default image used by the container
     REDIS_DEFAULT_IMAGE = "redis:latest"
 
-    attr_reader :port, :password
-
     # Initializes a new instance of RedisContainer
     #
     # @param image [String] the image to use
     # @param port [String] the port to use
     # @param kwargs [Hash] the options to pass to the container. See {DockerContainer#initialize}
     # @return [RedisContainer] a new instance of RedisContainer
-    def initialize(image = REDIS_DEFAULT_IMAGE, port: nil, **kwargs)
+    def initialize(image = REDIS_DEFAULT_IMAGE, **kwargs)
       super(image, **kwargs)
-      @port = port || ENV.fetch("REDIS_PORT", REDIS_DEFAULT_PORT)
-      @password = password || ENV.fetch("REDIS_PASSWORD", nil)
       @wait_for ||= add_wait_for(:logs, /Ready to accept connections/)
     end
 
@@ -32,18 +28,14 @@ module Testcontainers
     #
     # @return [RedisContainer] self
     def start
-      with_exposed_ports(@port)
-      _configure
+      with_exposed_ports(port)
       super
     end
 
-    # Sets the password to use
-    #
-    # @param password [String] the password to use
-    # @return [RedisContainer] self
-    def with_password(password)
-      @password = password
-      self
+    # Returns the port used by the container
+    # @return [Integer] the port used by the container
+    def port
+      REDIS_DEFAULT_PORT
     end
 
     # Returns the Redis connection url (e.g. redis://:password@localhost:6379/0)
@@ -61,11 +53,5 @@ module Testcontainers
     end
 
     alias_method :database_url, :redis_url
-
-    private
-
-    def _configure
-      add_env("REDIS_PASSWORD", @password) unless @password.nil? || @password.empty?
-    end
   end
 end