+++ redis/CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+### Fixed
+
+- Remove support for custom password and port, those aren't supported by
+  Redis images
+
 ## [0.1.0] - 2023-05-13
 
 ### Added